+++ src/main/java/net/weg/taskmanager/repository/OptionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OptionRepository extends JpaRepository<Text, Integer> {
+public interface OptionRepository extends JpaRepository<Text, Long> {
 }
