+++ src/main/java/net/weg/taskmanager/service/IService.java
@@ -4,9 +4,9 @@ import java.util.Collection;
 
 public interface IService <T> {
 
-    T findById(Integer id);
+    T findById(Long id);
     Collection<T> findAll();
-    void delete(Integer id);
+    void delete(Long id);
     T create(T obj);
     T update(T obj);
 }
